requirements.txt
backports-abc==0.4
backports.ssl-match-hostname==3.4.0.2
biopython==1.66
cvxopt==1.1.8
cycler==0.10.0
decorator==4.0.10
dill==0.2.5
eden==0.2
esmre==0.3.1
flake8==2.5.1
graphviz==0.4.10
hmmlearn==0.2.0
ipykernel==4.1.1
ipython==4.0.1
ipython-genutils==0.1.0
ipywidgets==4.1.0
Jinja2==2.8
joblib==0.9.4
jsonschema==2.4.0
jupyter==1.0.0
jupyter-client==4.1.1
jupyter-console==4.0.3
jupyter-core==4.0.6
MarkupSafe==0.23
matplotlib==1.5.1
mccabe==0.3.1
mistune==0.7.1
nbconvert==4.0.0
nbformat==4.0.1
networkx==1.10
nose==1.3.7
notebook==4.0.6
numpy==1.11.1
openbabel==1.8.3
pandas==0.18.1
path.py==0.0.0
pep8==1.5.7
pexpect==3.3
pickleshare==0.5
Pillow==3.2.0
ptyprocess==0.5
pybedtools==0.7.7
pycairo==1.10.0
pydot==1.0.2
pyflakes==1.0.0
Pygments==2.0.2
pygraphviz==1.2
PyMF==0.1.9
pyparsing==2.1.4
pysam==0.9.1
python-dateutil==2.5.3
pytz==2016.4
PyYAML==3.11
pyzmq==14.7.0
qtconsole==4.1.1
regex==2016.8.27
reportlab==3.3.0
requests==2.10.0
scikit-learn==0.17.1
scikit-neuralnetwork==0.4
scipy==0.17.1
scripttest==1.3
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
terminado==0.5
Theano==0.7.0
tornado==4.3
traitlets==4.0.0
weblogo==3.4
